bits and bobs
